XXXXService as an interface, old impl
Use interfaces when possible
