+++ server/src/main/java/org/obapanel/lockfactoryserver/server/service/semaphore/SemaphoreService.java
@@ -1,80 +1,34 @@
 package org.obapanel.lockfactoryserver.server.service.semaphore;
 
-import org.obapanel.lockfactoryserver.core.util.RuntimeInterruptedException;
-import org.obapanel.lockfactoryserver.server.LockFactoryConfiguration;
 import org.obapanel.lockfactoryserver.server.service.LockFactoryServices;
 import org.obapanel.lockfactoryserver.server.service.Services;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
-import java.util.concurrent.Semaphore;
 import java.util.concurrent.TimeUnit;
 
 /**
  * Service based on semaphore primitive
  */
-public class SemaphoreService implements LockFactoryServices {
+public interface SemaphoreService extends LockFactoryServices {
 
-    private static final Logger LOGGER = LoggerFactory.getLogger(SemaphoreService.class);
 
-    public static final Services TYPE = Services.SEMAPHORE;
+    Services TYPE = Services.SEMAPHORE;
 
-    private final SemaphoreCache semaphoreCache;
-
-    public SemaphoreService(LockFactoryConfiguration configuration) {
-        this.semaphoreCache = new SemaphoreCache(configuration);
-    }
-
-    public Services getType() {
+    default Services getType() {
         return TYPE;
     }
 
-    @Override
-    public void shutdown() throws Exception {
-        semaphoreCache.clearAndShutdown();
-    }
+    int currentPermits(String name);
 
-    public int currentPermits(String name) {
-        LOGGER.info("service> currentPermits name {}",name);
-        Semaphore semaphore = semaphoreCache.getOrCreateData(name);
-        return semaphore.availablePermits();
-    }
+    void acquire(String name, int permits);
 
+    boolean tryAcquire(String name, int permits);
 
-    public void acquire(String name, int permits) {
-        try {
-            LOGGER.info("service> acquire name{} permits {}", name, permits);
-            Semaphore semaphore = semaphoreCache.getOrCreateData(name);
-            semaphore.acquire(permits);
-        } catch (InterruptedException e) {
-            throw RuntimeInterruptedException.getToThrowWhenInterrupted(e);
-        }
-    }
-
-    public boolean tryAcquire(String name, int permits) {
-        LOGGER.info("service> tryAcquire name {} permits {}", name, permits);
-        Semaphore semaphore = semaphoreCache.getOrCreateData(name);
-        return semaphore.tryAcquire(permits);
-    }
-
-    public boolean tryAcquireWithTimeOut(String name, int permits, long timeOut) {
+    default boolean tryAcquireWithTimeOut(String name, int permits, long timeOut) {
         return this.tryAcquireWithTimeOut(name, permits, timeOut, TimeUnit.MILLISECONDS);
     }
 
-    public boolean tryAcquireWithTimeOut(String name, int permits, long timeOut, TimeUnit unit) {
-        try {
-            LOGGER.info("service> tryAcquireWithTimeOut name {} permits {} timeOut {} unit {}", name, permits, timeOut, unit);
-            Semaphore semaphore = semaphoreCache.getOrCreateData(name);
-            return semaphore.tryAcquire(permits, timeOut, unit);
-        } catch (InterruptedException e) {
-            throw RuntimeInterruptedException.getToThrowWhenInterrupted(e);
-        }
-    }
+    boolean tryAcquireWithTimeOut(String name, int permits, long timeOut, TimeUnit unit);
 
-    public void release(String name, int permits) {
-        LOGGER.info("service> release name {} permits {}", name, permits);
-        Semaphore semaphore = semaphoreCache.getOrCreateData(name);
-        semaphore.release(permits);
-    }
+    void release(String name, int permits);
 
 }
